Update tut8.java
Polymorphism in Java is one of the core concepts of Object-Oriented Programming (OOP). It refers to the ability of a single object to take on different forms.

Polymorphism can be categorized into two types:

1. Compile-time Polymorphism (Static Polymorphism):

This type of polymorphism is resolved at compile time. It is achieved through method overloading and operator overloading.

* Method Overloading: This occurs when multiple methods in the same class have the same name but differ in the number or type of parameters.

* Operator Overloading: Java does not support operator overloading

2. Runtime Polymorphism (Dynamic Polymorphism):

This type of polymorphism is resolved at runtime. It is achieved through method overriding and is based on inheritance. In runtime polymorphism, a subclass provides a specific implementation of a method that is already defined in its superclass.

Method Overriding: When a subclass provides its own implementation of a method that is already defined in its superclass, the subclass method is said to override the superclass method.

Key Points About Polymorphism:

Overloading happens at compile time (method overloading).
Overriding happens at runtime (method overriding).